fix: improve db code to use i64 types

+++ common/src/db_models/operating_system_version.rs
@@ -3,10 +3,10 @@ use utoipa::ToSchema;
 
 #[derive(Debug, Serialize, ToSchema)]
 pub struct OperatingSystemVersion {
-    pub id: i32,
+    pub id: i64,
     pub version: String,
-    pub operating_system_id: i32,
-    pub device_id: i32,
+    pub operating_system_id: i64,
+    pub device_id: i64,
 }
 
 impl From<entity::operating_system_version::Model> for OperatingSystemVersion {
