+++ entity/src/entities/executable_operating_system_version.rs
@@ -6,9 +6,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "executable_operating_system_version")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
-    pub executable_id: i32,
-    pub operating_system_version_id: i32,
+    pub id: i64,
+    pub executable_id: i64,
+    pub operating_system_version_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
