+++ entity/src/entities/operating_system_version.rs
@@ -6,10 +6,10 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "operating_system_version")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub version: String,
-    pub operating_system_id: i32,
-    pub device_id: i32,
+    pub operating_system_id: i64,
+    pub device_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
